rename: model 패키지의 내부 패키지 추가 후 이동
