+++ src/main/java/laddergame/model/people/Person.java
@@ -1,4 +1,4 @@
-package laddergame.model;
+package laddergame.model.people;
 
 import java.util.regex.Pattern;
