+++ src/test/java/laddergame/model/ladder/HeightTest.java
@@ -1,4 +1,4 @@
-package laddergame.model;
+package laddergame.model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
