+++ src/test/java/laddergame/model/ladder/LineTest.java
@@ -1,4 +1,4 @@
-package laddergame.model;
+package laddergame.model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
